+++ example/lib/main.dart
@@ -190,6 +190,7 @@ class _MyAppState extends State<MyApp> {
       case ReorderableType.gridViewBuilder:
         return ReorderableBuilder.builder(
           key: Key(_gridViewKey.toString()),
+          positionDuration: const Duration(seconds: 1),
           onReorder: _handleReorder,
           lockedIndices: lockedIndices,
           nonDraggableIndices: nonDraggableIndices,